pyproject.toml
[project]
name = "markdocx"
version = "0.1.0"
description = "Convert AI-generated Markdown textbooks to polished DOCX with native math equations and syntax-highlighted code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "latex2mathml==3.78.1",
    "lxml==6.0.2",
    "markdown-it-py==4.0.0",
    "matplotlib==3.10.8",
    "mdit-py-plugins==0.5.0",
    "pillow==12.1.0",
    "pygments==2.19.2",
    "python-docx==1.2.0",
]

[project.scripts]
markdocx = "markdocx.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/markdocx"]
